// src/server/db/schema.generated.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface ChatMessages {
  chat_id: number;
  content: string;
  created_at: Generated<Timestamp>;
  id: Generated<number>;
  role: string;
}

export interface Chats {
  created_at: Generated<Timestamp>;
  id: Generated<number>;
  title: string | null;
  updated_at: Generated<Timestamp>;
  user_id: string;
}

export interface Documents {
  content: string;
  created_at: Generated<Timestamp>;
  id: Generated<number>;
  title: string;
  updated_at: Generated<Timestamp>;
  user_id: string;
}

export interface Users {
  created_at: Generated<Timestamp>;
  id: string;
  name: string;
}

export interface DB {
  chat_messages: ChatMessages;
  chats: Chats;
  documents: Documents;
  users: Users;
}